chore(gb-9270): simplify router creation
