+++ crates/federated-server/src/server/engine_reloader.rs
@@ -46,20 +46,18 @@ impl GatewayEngineReloader {
     /// 1. Waits for the initial graph definition
     /// 2. Builds the initial engine
     /// 3. Spawns a background task that listens for updates and rebuilds the engine
-    pub async fn spawn(config: EngineReloaderConfig<'_>) -> crate::Result<Self> {
-        // Destructure the config to get all fields
-        let EngineReloaderConfig {
+    pub async fn spawn(
+        EngineReloaderConfig {
             mut update_receiver,
             initial_config,
             extension_catalog,
             logging_filter,
             hot_reload_config_path,
             access_token,
-        } = config;
-
+        }: EngineReloaderConfig<'_>,
+    ) -> crate::Result<Self> {
         let mut current_config = initial_config;
 
-        // Wait for the initial graph definition
         tracing::debug!("Waiting for the initial graph...");
 
         let mut graph_definition = loop {
@@ -92,7 +90,6 @@ impl GatewayEngineReloader {
         let engine = build_engine(initial_context, graph_definition.clone(), vec![]).await?;
         let (engine_sender, engine_watcher) = watch::channel(engine);
 
-        // Spawn the update loop as a background task
         tokio::spawn(async move {
             let mut in_progress_reload: Option<JoinHandle<()>> = None;
 
@@ -102,13 +99,11 @@ impl GatewayEngineReloader {
                     reload.abort();
                 }
 
-                // Update our state based on the event type
                 match update {
                     UpdateEvent::Graph(new_graph) => graph_definition = new_graph,
                     UpdateEvent::Config(new_config) => current_config = *new_config,
                 }
 
-                // Spawn a new task to build the updated engine
                 in_progress_reload = Some(tokio::spawn({
                     let hot_reload_config_path = hot_reload_config_path.clone();
                     let access_token = access_token.clone();
@@ -118,10 +113,8 @@ impl GatewayEngineReloader {
                     let logging_filter = logging_filter.clone();
 
                     async move {
-                        // Extract operations to warm from the current engine
                         let operations_to_warm = extract_operations_to_warm(&current_config, &engine_sender);
 
-                        // Build the context for the new engine
                         let context = EngineBuildContext {
                             gateway_config: &current_config,
                             hot_reload_config_path: hot_reload_config_path.as_ref(),
@@ -130,7 +123,6 @@ impl GatewayEngineReloader {
                             logging_filter: &logging_filter,
                         };
 
-                        // Build the new engine
                         match build_engine(context, graph_definition, operations_to_warm).await {
                             Ok(new_engine) => {
                                 if let Err(err) = engine_sender.send(new_engine) {
