+++ crates/federated-server/src/server/router/health.rs
